refactor: rearranged file structures - divided design.ts into two files - created a new file for game rules
